Module cgl_rs::graphics

The CGL Graphics Module

Modules

text

The text module contains the text rendering, font loading functionality of CGL

widgets

The CGL Widgets Module. This is powered by th epowerful cgl widgets batch renderer.

Structs

Bloom

The Bloom Object

Framebuffer

A framebuffer object that can be used for offscreen rendering.

MeshCPU

MeshCPU_C

The Mesh Structure

MeshGPU

MeshVertex

The Mesh Vertex Structure

Shader

A shader object

ShaderStorageBufferObject

The Shader Storage Buffer Object

Texture

The texture object

Tilemap

The Tilemap Object

Enums

TextureDataType

The texture data type used by CGL These are reflections of OpenGL’s enum values You can also use the _i version of methods to pass integers directly instead of enums

TextureFormat

The texture format used by CGL These are reflections of OpenGL’s enum values You can also use the _i version of methods to pass integers directly instead of enums

TextureInternalFormat

The texture internal format used by CGL These are reflections of OpenGL’s enum values You can also use the _i version of methods to pass integers directly instead of enums

TextureScalingMode

The texture scaling mode used by CGL These are reflections of OpenGL’s enum values You can also use the _i version of methods to pass integers directly instead of enums

TextureWrappingMode

The texture wrapping mode used by CGL These are reflections of OpenGL’s enum values You can also use the _i version of methods to pass integers directly instead of enums

Functions

clear

Clears the screen with the specified color.

init

Initializes the CGL Graphics Module.

render_screen_quad

Renders a screen quad.

shutdown

Shuts down the CGL Graphics Module.

Type Definitions

SSBO