Module cgl_rs::graphics::widgets

The CGL Widgets Module. This is powered by th epowerful cgl widgets batch renderer.

Functions

add_arc2f

Adds an arc to the current frame.

add_character

Adds a character to the current frame.

add_circle

Adds a circle to the current frame.

add_circle2f

Adds a circle to the current frame.

add_circle2fr

Adds a circle to the current frame.

add_circler

Adds a circle to the current frame.

add_cubic_bazier

Adds a cubic bezier curve to the current frame.

add_cubic_bazier2f

Adds a cubic bezier curve to the current frame.

add_cubic_bazier2v

Adds a cubic bezier curve to the current frame.

add_cubic_bazier_points

Adds a cubic bezier curve to the current frame using points.

add_cubic_bazier_points2v

Arguments

add_line

Adds a line to the current frame.

add_line2f

Adds a line to the current frame.

add_oval

Adds an oval to the current frame.

add_oval2f

Adds an oval to the current frame.

add_oval2fr

Adds an oval to the current frame.

add_quad

Adds a quad to the current frame.

add_quad_8f

Adds a quad to the current frame.

add_rect

Adds a rectangle to the current frame.

add_rect2f

Adds a rectangle to the current frame.

add_rounded_rect2f

Adds a rounded rectangle to the current frame.

add_string

Adds a string to the current frame.

add_string_with_font

Adds a string to the current frame with a specified font.

add_triangle

Adds a triangle to the current frame.

add_vertex_p

Adds a vertex with the given position to the current batch.

add_vertex_p3f

Adds a vertex with the given position to the current batch.

add_vertex_p3ft

Adds a vertex with the given position and texture coordinate to the current batch.

add_vertex_p3ft2f

Adds a vertex with the given position and texture coordinate to the current batch.

add_vertex_pt

Adds a vertex with the given position and texture coordinate to the current batch.

add_vertex_pt2f

Adds a vertex with the given position and texture coordinate to the current batch.

apply_transformations_on_cpu

This enables the transformation application on the CPU.

apply_transformations_on_gpu

This enables the transformation application on the GPU.

begin

Begins a new CGL frame

begin_int

Begins a new CGL frame with a global scale and offset.

disable_diffuse_shading

Disables diffuse shading for the current frame.

enable_diffuse_shading

Enables diffuse shading for the current frame.

end

Ends the current CGL frame

flush

Flushes all the vertices added to the current batch.

flush_if_required

Flushes all the vertices added to the current batch if required.

init

Initializes the CGL widgets

set_fill_color

Sets the fill color for the current frame.

set_fill_colorf

Sets the fill color for the current frame.

set_fill_mode

Sets the fill mode for the current frame.

set_font_texture

Sets the font texture for the current frame.

set_mask

Sets the mask rendering mask for the current frame.

set_maskf

Sets the mask rendering mask for the current frame.

set_model_matrix

Sets the model matrix for the current frame.

set_projection_matrix

Sets the projection matrix for the current frame.

set_stroke_color

Sets the stroke color for the current frame.

set_stroke_colorf

Sets the stroke color for the current frame.

set_stroke_thickness

Sets the stroke thickness for the current frame.

set_texture

Sets the texture for the current frame.

set_texture_coordinate_so

Sets the texture coordinate scale and offset for the current frame.

set_view_matrix

Sets the view matrix for the current frame.

shutdown

Shuts down the CGL widgets