Struct cgl_rs::math::Matrix3x3

#[repr(C)]
pub struct Matrix3x3 { /* private fields */ }

A 3x3 matrix struct used for graphics programming with OpenGL.

Implementations

impl Matrix3x3

pub fn new( m0: f32, m1: f32, m2: f32, m3: f32, m4: f32, m5: f32, m6: f32, m7: f32, m8: f32 ) -> Matrix3x3

Creates a new 3x3 matrix.

Arguments

m0 … m8: The values of the matrix.

Returns

A new 3x3 matrix.

Example

use cgl_rs::math::Matrix3x3;
 
let m = Matrix3x3::new(
   1.0, 0.0, 0.0,
   0.0, 1.0, 0.0,
   0.0, 0.0, 1.0
);

pub fn zero() -> Matrix3x3

Creates a new 3x3 matrix with all elements set to zero.

Returns

A new 3x3 matrix with all elements set to zero.

Example

use cgl_rs::math::Matrix3x3;

let m = Matrix3x3::zero();

pub fn identity() -> Matrix3x3

Creates a new 3x3 identity matrix.

Returns

A new 3x3 identity matrix.

Example

use cgl_rs::math::Matrix3x3;

let m = Matrix3x3::identity();

pub fn transpose(&self) -> Matrix3x3

Transposes the matrix.

Returns

A new 3x3 matrix that is the transpose of the original matrix.

Example

use cgl_rs::math::Matrix3x3;

let m = Matrix3x3::new(
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0
);

let m_transpose = m.transpose();

pub fn trace(&self) -> f32

Calculates the trace of the matrix.

Returns

The sum of the diagonal elements of the matrix.

Example

use cgl_rs::math::Matrix3x3;

let m = Matrix3x3::new(
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0
);

let trace = m.trace();

pub fn determinant(&self) -> f32

Calculates the determinant of the matrix.

Returns

The determinant of the matrix.

Example

use cgl_rs::math::Matrix3x3;

let m = Matrix3x3::new(
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0
);

let det = m.determinant();