Struct cgl_rs::math::Matrix4x4

#[repr(C)]
pub struct Matrix4x4 { /* private fields */ }

A 4x4 matrix struct used for graphics programming with OpenGL.

Implementations

impl Matrix4x4

pub fn new( m00: f32, m01: f32, m02: f32, m03: f32, m10: f32, m11: f32, m12: f32, m13: f32, m20: f32, m21: f32, m22: f32, m23: f32, m30: f32, m31: f32, m32: f32, m33: f32 ) -> Matrix4x4

Creates a new matrix with the given values.

Arguments

m00 .. m33 - The values of the matrix.

Returns

A new identity matrix.

Example

use cgl_rs::math::Matrix4x4;
 
let m = Matrix4x4::new(
   1.0, 0.0, 0.0, 0.0,
   0.0, 1.0, 0.0, 0.0,
   0.0, 0.0, 1.0, 0.0,
   0.0, 0.0, 0.0, 1.0
);
 
println!("{}", m);

pub fn from_mat3(m: &Matrix3x3) -> Matrix4x4

Creates a new 4x4 matrix from a 3x3 matrix.

Arguments

m - The 3x3 matrix to convert to a 4x4 matrix.

Returns

A new 4x4 matrix with the same values as the input 3x3 matrix, with the fourth row and column set to zero.

Example

use cgl_rs::math::{Matrix3x3, Matrix4x4};

let m3 = Matrix3x3::new(
    1.0, 2.0, 3.0,
    4.0, 5.0, 6.0,
    7.0, 8.0, 9.0
);

let m4 = Matrix4x4::from_mat3(&m3);

pub fn zero() -> Matrix4x4

Returns a new matrix with all elements set to zero.

Returns

A new matrix with all elements set to zero.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::zero();

pub fn identity() -> Matrix4x4

Returns a new identity matrix.

Returns

A new identity matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();

pub fn scale(x: f32, y: f32, z: f32) -> Matrix4x4

Returns a new scaling matrix.

Arguments

x - The scaling factor along the x-axis.

y - The scaling factor along the y-axis.

z - The scaling factor along the z-axis.

Returns

A new scaling matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::scale(2.0, 3.0, 4.0);

pub fn translate(x: f32, y: f32, z: f32) -> Matrix4x4

Returns a new translation matrix.

Arguments

x - The translation along the x-axis.

y - The translation along the y-axis.

z - The translation along the z-axis.

Returns

A new translation matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::translate(1.0, 2.0, 3.0);

pub fn rotate_x(angle: f32) -> Matrix4x4

Returns a new rotation matrix around the x-axis.

Arguments

angle - The angle of rotation in radians.

Returns

A new rotation matrix around the x-axis.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::rotate_x(0.5);

pub fn rotate_y(angle: f32) -> Matrix4x4

Returns a new rotation matrix around the y-axis.

Arguments

angle - The angle of rotation in radians.

Returns

A new rotation matrix around the y-axis.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::rotate_y(0.5);

pub fn rotate_z(angle: f32) -> Matrix4x4

Returns a new rotation matrix around the z-axis.

Arguments

angle - The angle of rotation in radians.

Returns

A new rotation matrix around the z-axis.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::rotate_z(0.5);

pub fn rotate_about_axis(axis: &Vector3, angle: f32) -> Matrix4x4

Returns a new rotation matrix that rotates around the given axis by the specified angle.

Arguments

axis - The axis to rotate around.

angle - The angle of rotation in radians.

Returns

A new rotation matrix that rotates around the given axis by the specified angle.

Example

use cgl_rs::math::{Matrix4x4, Vector3};

let axis = Vector3::new(1.0, 0.0, 0.0);
let m = Matrix4x4::rotate_about_axis(&axis, 0.5);

pub fn perspective(fov: f32, aspect: f32, near: f32, far: f32) -> Matrix4x4

Returns a new perspective projection matrix.

Arguments

fov - The field of view angle in radians.

aspect - The aspect ratio of the projection.

near - The distance to the near clipping plane.

far - The distance to the far clipping plane.

Returns

A new perspective projection matrix.

Example

use cgl_rs::math::Matrix4x4;
 
let m = Matrix4x4::perspective(cgl_rs::math::constants::PI_2, 16.0/9.0, 0.1, 100.0);

pub fn orthographic( left: f32, right: f32, bottom: f32, top: f32, near: f32, far: f32 ) -> Matrix4x4

Returns a new orthographic projection matrix.

Arguments

left - The coordinate for the left vertical clipping plane.

right - The coordinate for the right vertical clipping plane.

bottom - The coordinate for the bottom horizontal clipping plane.

top - The coordinate for the top horizontal clipping plane.

near - The distance to the near clipping plane.

far - The distance to the far clipping plane.

Returns

A new orthographic projection matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::orthographic(-1.0, 1.0, -1.0, 1.0, 0.1, 100.0);

pub fn look_at(eye: Vector3, target: Vector3, up: Vector3) -> Matrix4x4

Returns a new view matrix that looks from the eye position towards the target position.

Arguments

eye - The position of the camera.

target - The position to look at.

up - The up direction of the camera.

Returns

A new view matrix.

Example

use cgl_rs::math::{Matrix4x4, Vector3};

let mat = Matrix4x4::look_at(
   Vector3::new(0.0, 0.0, 0.0),
   Vector3::new(0.0, 0.0, -1.0),
   Vector3::new(0.0, 1.0, 0.0)
);

pub fn mul(&self, other: &Matrix4x4) -> Matrix4x4

Multiplies this matrix by another matrix and returns the result.

Arguments

other - The matrix to multiply by.

Returns

The resulting matrix of the multiplication.

Example

use cgl_rs::math::Matrix4x4;

let m1 = Matrix4x4::identity();
let m2 = Matrix4x4::rotate_z(0.5);
let m3 = m1.mul(&m2);

pub fn determinant(&self) -> f32

Calculates the determinant of this matrix.

Returns

The determinant of this matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let det = m.determinant();

pub fn mul_vec4(&self, other: &Vector4) -> Vector4

Multiplies this matrix by a vector and returns the result.

Arguments

other - The vector to multiply by.

Returns

The resulting vector of the multiplication.

Example

use cgl_rs::math::{Matrix4x4, Vector4};

let m = Matrix4x4::identity();
let v = Vector4::new(1.0, 2.0, 3.0, 1.0);
let result = m.mul_vec4(&v);

pub fn inverse(&self) -> Matrix4x4

Calculates the inverse of this matrix.

Returns

The inverse of this matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let inv = m.inverse();

pub fn transpose(&self) -> Matrix4x4

Transposes this matrix.

Returns

The transposed matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let transposed = m.transpose();

pub fn adjoint(&self) -> Matrix4x4

Calculates the adjoint of this matrix.

Returns

The adjoint of this matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let adj = m.adjoint();

pub fn gaussian_elimination(&self) -> Matrix4x4

Performs Gaussian elimination on this matrix.

Returns

The matrix in row echelon form.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let row_echelon = m.gaussian_elimination();

pub fn rank(&self) -> i32

Calculates the rank of this matrix.

Returns

The rank of this matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let rank = m.rank();

pub fn trace(&self) -> f32

Calculates the trace of this matrix.

Returns

The trace of this matrix.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let trace = m.trace();

pub fn to_mat3(&self) -> Matrix3x3

Converts this Matrix4x4 to a Matrix3x3 by discarding the last row and column.

Example

use cgl_rs::math::{Matrix3x3, Matrix4x4};

let m = Matrix4x4::identity();
let m3 = m.to_mat3();

pub fn decompose_lu(&self) -> (Matrix4x4, Matrix4x4)

Decomposes this matrix into lower and upper triangular matrices using LU decomposition.

Returns

A tuple containing the lower and upper triangular matrices.

Example

use cgl_rs::math::Matrix4x4;

let m = Matrix4x4::identity();
let (l, u) = m.decompose_lu();

pub fn lerp(&self, other: &Matrix4x4, t: f32) -> Matrix4x4

Performs a linear interpolation between this matrix and another matrix.

Arguments

other - The other matrix to interpolate with.

t - The interpolation factor. Should be between 0.0 and 1.0.

Returns

The interpolated matrix.

Example

use cgl_rs::math::Matrix4x4;

let m1 = Matrix4x4::identity();
let m2 = Matrix4x4::scale(1.0, 2.0, 3.0);
let m3 = m1.lerp(&m2, 0.5);