Struct cgl_rs::math::Vector4

#[repr(C)]
pub struct Vector4 {
    pub x: f32,
    pub y: f32,
    pub z: f32,
    pub w: f32,
}

A 4-dimensional vector with x, y, z and w components.

Fields
x: f32y: f32z: f32w: f32
Implementations

impl Vector4

pub fn new(x: f32, y: f32, z: f32, w: f32) -> Vector4

Creates a new Vector4 with the given x, y, z, and w components.

Arguments

x - The x component of the new Vector4.

y - The y component of the new Vector4.

z - The z component of the new Vector4.

w - The w component of the new Vector4.

Examples

use cgl_rs::math::Vector4;

let vec = Vector4::new(1.0, 2.0, 3.0, 4.0);
assert_eq!(vec.x, 1.0);
assert_eq!(vec.y, 2.0);
assert_eq!(vec.z, 3.0);
assert_eq!(vec.w, 4.0);

pub fn from_vec3(vec: Vector3, w: f32) -> Vector4

Creates a new Vector4 from a Vector3 and a w component.

Arguments

vec - The Vector3 to use as the x, y, and z components of the new Vector4.

w - The w component of the new Vector4.

Examples

use cgl_rs::math::{Vector3, Vector4};

let vec3 = Vector3::new(1.0, 2.0, 3.0);
let vec4 = Vector4::from_vec3(vec3, 4.0);
assert_eq!(vec4.x, 1.0);
assert_eq!(vec4.y, 2.0);
assert_eq!(vec4.z, 3.0);
assert_eq!(vec4.w, 4.0);

pub fn from_vec2(vec: Vector2, z: f32, w: f32) -> Vector4

Creates a new Vector4 from a Vector2, z, and w components.

Arguments

vec - The Vector2 to use as the x and y components of the new Vector4.

z - The z component of the new Vector4.

w - The w component of the new Vector4.

Examples

use cgl_rs::math::{Vector2, Vector4};

let vec2 = Vector2::new(1.0, 2.0);
let vec4 = Vector4::from_vec2(vec2, 3.0, 4.0);
assert_eq!(vec4.x, 1.0);
assert_eq!(vec4.y, 2.0);
assert_eq!(vec4.z, 3.0);
assert_eq!(vec4.w, 4.0);

pub fn zero() -> Vector4

Returns a new Vector4 with all components set to 0.0.

Examples

use cgl_rs::math::Vector4;

let vec = Vector4::zero();
assert_eq!(vec.x, 0.0);
assert_eq!(vec.y, 0.0);
assert_eq!(vec.z, 0.0);
assert_eq!(vec.w, 0.0);

pub fn one() -> Vector4

Returns a new Vector4 with all components set to 1.0.

Examples

use cgl_rs::math::Vector4;

let vec = Vector4::one();
assert_eq!(vec.x, 1.0);
assert_eq!(vec.y, 1.0);
assert_eq!(vec.z, 1.0);
assert_eq!(vec.w, 1.0);

pub fn xyz(&self) -> Vector3

Extracts the x, y, and z components of the Vector4 as a new Vector3.

Examples

use cgl_rs::math::{Vector3, Vector4};

let vec4 = Vector4::new(1.0, 2.0, 3.0, 4.0);
let vec3 = vec4.xyz();
assert_eq!(vec3.x, 1.0);
assert_eq!(vec3.y, 2.0);
assert_eq!(vec3.z, 3.0);