Module cgl_rs::utils

This module contains CGL utility functions, as well some miscellaneous functions.

Functions

append_file

Arguments

append_string_to_file

Appends a string to a file.

big_endian_to_current

Converts a slice of u8 values from big-endian byte order to the system’s native byte order.

crc32

Calculates the CRC32 checksum of the given data.

crc64

The C implementation is based on linux’s crc64 implementation.

fast_rand

Generates a fast random number.

fast_srand

Sets the seed for the fast random number generator.

get_file_size

Returns the size of a file in bytes.

get_random_with_probability

Returns a random index from the given slice of probabilities, where the probability of each index being chosen is proportional to its value in the slice.

get_time

Returns the time in seconds since the program started.

get_timestamp

Returns a string representing the current timestamp.

is_little_endian

Returns a boolean indicating whether the system is little-endian or not.

lerp

Linearly interpolates between two values.

little_endian_to_current

Converts a slice of u8 values from little-endian byte order to the system’s native byte order.

map

Maps a value from one range to another.

rand31

Generates a random number using the rand31 algorithm.

random_float

Generates a random floating-point number between 0 and 1.

random_float_range

Generates a random floating-point number within the given range.

random_int

Generates a random integer within the given range.

read_file

Reads the contents of a file and returns it as a vector of bytes.

read_file_as_string

Reads the contents of a file and returns it as a string.

reverse_bytes

Reverses the order of bytes in the given slice of u8 values.

rot13

Applies the ROT13 substitution cipher to the given string.

sleep

Suspends the current thread for the specified amount of time.

srand31

Sets the seed for the rand31 random number generator.

super_fast_hash

Calculates the SuperFastHash of the given data.

write_file

Arguments

write_string_to_file

Writes a string to a file.

xorshf96

Generates a fast random number using the xorshf96 algorithm.